fix(bundle): Creating a bundle for the lib
So far, the library was built using tsc (so that it could be tree-shaked).
But because of ASSETS_URL (an environment variable that must be replaced at build time by Webpack), tsc is not sufficient.
Indeed, the process.env reference is left by tsc.

So this commit builds a single non-shakable package "index.js" to be used by the library authors. the bundle.js remains. It is index.js + side-effects (the booting of the lib)