Getting rid of the tutorial
Due to a bug, the "old" tutorial has not been reachable from WorkAdventure for quite some time.
The "init" method of the tutorial was not fired when booting the scripting API extra.

Because this tutorial is old and deprecated, and since no one complained it was not working, I assume nobody is using it.
I therefore decide to completely get rid of it.

Bonus: we can get rid of the Nes.CSS dependency that causes issues when installing packages.